refactor: plugin-factory API + remove getDevServer bridge
Replace the globalThis dev-server bridge with an explicit createApp({
viteDevServer }) factory; the Vite plugin passes the dev server directly
and production gates listen() on an ESM main-module check.

Make reactRouterFastify a plugin factory — app.register(reactRouterFastify({
... })) — so getLoadContext is typed for your server. Flatten src/ to
server.ts / plugin.ts / vite.ts / index.ts and trim exports to "." and
"./vite".
