+++ packages/remix-fastify/src/server.ts
@@ -14,9 +14,10 @@ import type {
 import type {
   AppLoadContext,
   RouterContextProvider,
+  ServerBuild,
   UNSAFE_MiddlewareEnabled as MiddlewareEnabled,
 } from "react-router";
-import type { ViteDevServer } from "vite";
+import { createRequestHandler as createReactRouterRequestHandlerImpl } from "react-router";
 
 type MaybePromise<T> = T | Promise<T>;
 
@@ -29,7 +30,7 @@ export type HttpServer =
 export type HttpRequest = RawRequestDefaultExpression<HttpServer>;
 export type HttpResponse = RawReplyDefaultExpression<HttpServer>;
 
-export type RequestHandler<Server extends HttpServer> = (
+export type RequestHandler<Server extends HttpServer = HttpServer> = (
   request: FastifyRequest<RouteGenericInterface, Server>,
   reply: FastifyReply<RouteGenericInterface, Server>,
 ) => Promise<void>;
@@ -39,17 +40,43 @@ export type RequestHandler<Server extends HttpServer> = (
  * `action` functions.
  *
  * You can think of this as an escape hatch that allows you to pass
- * environment/platform-specific values through to your loader/action.
+ * environment/platform-specific values through to your loader/action, such as
+ * values generated by Fastify hooks/decorators (e.g. `request.session`).
  *
  * When the `v8_middleware` future flag is enabled this must return a
  * `RouterContextProvider`; otherwise it returns your augmented `AppLoadContext`.
  */
-export type GetLoadContextFunction<Server extends HttpServer> = (
+export type GetLoadContextFunction<Server extends HttpServer = HttpServer> = (
   request: FastifyRequest<RouteGenericInterface, Server>,
   reply: FastifyReply<RouteGenericInterface, Server>,
 ) => MiddlewareEnabled extends true
   ? MaybePromise<RouterContextProvider>
   : MaybePromise<AppLoadContext>;
+
+/**
+ * Returns a request handler for Fastify that serves the response using React
+ * Router.
+ */
+export function createReactRouterRequestHandler<
+  Server extends HttpServer = HttpServer,
+>({
+  build,
+  getLoadContext,
+  mode = process.env.NODE_ENV,
+}: {
+  build: ServerBuild | (() => ServerBuild | Promise<ServerBuild>);
+  getLoadContext?: GetLoadContextFunction<Server>;
+  mode?: string;
+}): RequestHandler<Server> {
+  let handleRequest = createReactRouterRequestHandlerImpl(build, mode);
+
+  return async (request, reply) => {
+    let request_ = createReactRouterRequest(request, reply);
+    let loadContext = await getLoadContext?.(request, reply);
+    let response = await handleRequest(request_, loadContext);
+    return sendResponse(reply, response);
+  };
+}
 
 export function createHeaders(
   requestHeaders: FastifyRequest["headers"],
@@ -75,12 +102,12 @@ export function getUrl<Server extends HttpServer>(
   request: FastifyRequest<RouteGenericInterface, Server>,
 ): string {
   let origin = `${request.protocol}://${request.host}`;
-  // Use `request.originalUrl` so Remix and React Router are aware of the full path
+  // Use `request.originalUrl` so React Router is aware of the full path
   let url = `${origin}${request.originalUrl}`;
   return url;
 }
 
-export function createRequest<Server extends HttpServer>(
+export function createReactRouterRequest<Server extends HttpServer>(
   request: FastifyRequest<RouteGenericInterface, Server>,
   reply: FastifyReply<RouteGenericInterface, Server>,
 ): Request {
@@ -125,32 +152,6 @@ export async function sendResponse<Server extends HttpServer>(
   }
 
   return reply.send(await nodeResponse.text());
-}
-
-const DEV_SERVER_KEY = "__mcansh_remix_fastify_vite_dev_server__";
-
-/**
- * Stashes the Vite dev server on `globalThis` so that {@link reactRouterFastify}
- * (which runs inside your server entry, loaded via `ssrLoadModule`) can read it
- * without a direct reference. Set by the `fastifyDevServer` Vite plugin.
- */
-export function setDevServer(server: ViteDevServer): void {
-  (globalThis as Record<string, unknown>)[DEV_SERVER_KEY] = server;
-}
-
-/**
- * Returns the Vite dev server when running under `fastifyDevServer`, otherwise
- * `undefined`. Use this in your server entry to only call `app.listen()` in
- * production:
- *
- * ```ts
- * if (!getDevServer()) await app.listen({ port: 3000 });
- * ```
- */
-export function getDevServer(): ViteDevServer | undefined {
-  return (globalThis as Record<string, unknown>)[DEV_SERVER_KEY] as
-    | ViteDevServer
-    | undefined;
 }
 
 function responseToReadable(response: Response): Readable | null {
